Renamed Thread Border Router Management cluster files to PascalCase
Created compat stubs for the old headers to preserve compatibility.

Updated app_config_dependent_sources.gni and app_config_dependent_sources.cmake.

Tested by building linux-x64-all-devices.
