+++ src/app/clusters/thread-border-router-management-server/app_config_dependent_sources.gni
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "thread-border-router-management-server.cpp",
-  "thread-border-router-management-server.h",
-  "thread-br-delegate.h",
+  "ThreadBorderRouterManagementCluster.cpp",
+  "ThreadBorderRouterManagementCluster.h",
+  "ThreadBorderRouterManagementDelegate.h",
 ]
